chore: move cli scripts
