+++ pyproject.toml
@@ -40,8 +40,8 @@ dependencies = [
   "numba>=0.65",
 ]
 [project.scripts]
-om_init_hpc = "organelle_morphology.hpc_init:main"
-om_app = "organelle_morphology.om_app:run_marimo_app"
+om_init_hpc = "organelle_morphology.cli.hpc_init:main"
+om_app = "organelle_morphology.cli.om_app:run_marimo_app"
 
 [dependency-groups]
 dev = [
